+++ CHANGELOG.md
@@ -55,6 +55,8 @@ project.
   All basic statistics are now calculated for the error, relative error, absolute error and the relative absolute error.
   If `pingouin` is installed, we also calculte the ICC.
   (https://github.com/mad-lab-fau/gaitmap/pull/13)
+- A new utility method to intersect stride lists by a region of interest list.
+  ()
 
 ### Changed
 - When using `memory` with any of Dtw methods, we will not cache the peak detection step anymore.
@@ -69,6 +71,15 @@ project.
   predicted.
   This change requires updating the function call in your code.
   (https://github.com/mad-lab-fau/gaitmap/pull/13)
+- For all trajectory/pos/ori methods, the `sampling_rate_hz` parameter is now keyword only.
+  This might require some user facing code changes.
+  ()
+- Trajectory/pos/ori and ZUPT methods can now get an optional parameter called `stride_event_list` as parameter for the detect 
+  method.
+  So far, only the RTS Kalman Filter is using this parameter to forward it to the ZUPT method.
+  Both wrapper (Stride-level and Region) are both aware of this parameter and pass them correctly down to the respective
+  methods.
+  ()
 
 ### Removed
 - The `find_zupts` method of the `RtsKalmanFilter` and all deprecated arguments are now fully removed in favor of the